IRIS-3881 | moved logintests to auth
