+++ src/test/java/com/wikia/webdriver/testsuites/testSuite.xml
@@ -132,9 +132,9 @@
             <class name="com.wikia.webdriver.testcases.interactivemapstests.PalantirTests"/>
             <class name="com.wikia.webdriver.testcases.interactivemapstests.PinMapTests"/>
             <class name="com.wikia.webdriver.testcases.interactivemapstests.PinTypeMapTests"/>
-            <class name="com.wikia.webdriver.testcases.logintests.ForcedLoginTests"/>
-            <class name="com.wikia.webdriver.testcases.logintests.ForgottenPasswordTests"/>
-            <class name="com.wikia.webdriver.testcases.logintests.LoginTests"/>
+            <class name="com.wikia.webdriver.testcases.auth.ForcedLoginTests"/>
+            <class name="com.wikia.webdriver.testcases.auth.ForgottenPasswordTests"/>
+            <class name="com.wikia.webdriver.testcases.auth.LoginTests"/>
             <class name="com.wikia.webdriver.testcases.mediatests.addvideo.VetArticleCommentsTests"/>
             <class name="com.wikia.webdriver.testcases.mediatests.addvideo.VetArticlePlaceholderTests"/>
             <class name="com.wikia.webdriver.testcases.mediatests.addvideo.VetSpecialVideosTests"/>
